Refactor scheduler into routed screens and nested stacks
- Split monolithic scheduler page into dedicated overview, calendar, tasks, alerts, and task detail screens
- Add nested Expo Router layouts/routes for scheduler sections and task CRUD flows
- Remove sheet-based task create/detail UI in favor of full-screen routes
- Update tabs: fix Finance label and use calendar icon for Scheduler
