+++ apps/native/app/(tabs)/scheduler/[id]/_layout.tsx
@@ -2,10 +2,10 @@ import { Stack } from "expo-router";
 
 export default function SchedulerDetailLayout() {
   return (
-    <Stack
-      screenOptions={{
-        headerShown: false,
-      }}
-    />
+    <Stack>
+      <Stack.Screen name="index" options={{ title: "Task Detail" }} />
+      <Stack.Screen name="update" options={{ title: "Edit Task" }} />
+      <Stack.Screen name="delete" options={{ title: "Delete Task" }} />
+    </Stack>
   );
 }
