+++ apps/native/app/(tabs)/scheduler/alerts/_layout.tsx
@@ -1,0 +1,9 @@
+import { Stack } from "expo-router";
+
+export default function SchedulerAlertsLayout() {
+  return (
+    <Stack>
+      <Stack.Screen name="index" options={{ title: "Alerts" }} />
+    </Stack>
+  );
+}
