+++ apps/native/app/(tabs)/scheduler/alerts/index.tsx
@@ -1,0 +1,5 @@
+import { SchedulerAlertsScreen } from "@/components/scheduler/scheduler-alerts-screen";
+
+export default function SchedulerAlertsRoute() {
+  return <SchedulerAlertsScreen />;
+}
