+++ apps/native/app/(tabs)/scheduler/calendar/_layout.tsx
@@ -1,0 +1,9 @@
+import { Stack } from "expo-router";
+
+export default function SchedulerCalendarLayout() {
+  return (
+    <Stack>
+      <Stack.Screen name="index" options={{ title: "Calendar" }} />
+    </Stack>
+  );
+}
