+++ apps/native/app/(tabs)/scheduler/calendar/index.tsx
@@ -1,0 +1,5 @@
+import { SchedulerCalendarScreen } from "@/components/scheduler/scheduler-calendar-screen";
+
+export default function SchedulerCalendarRoute() {
+  return <SchedulerCalendarScreen />;
+}
