--- apps/native/app/(tabs)/scheduler/create.tsx
@@ -1,5 +1,0 @@
-import SchedulerScreen from "./index";
-
-export default function SchedulerCreateRoute() {
-  return <SchedulerScreen />;
-}
