+++ apps/native/app/(tabs)/scheduler/tasks/_layout.tsx
@@ -1,0 +1,10 @@
+import { Stack } from "expo-router";
+
+export default function SchedulerTasksLayout() {
+  return (
+    <Stack>
+      <Stack.Screen name="index" options={{ title: "Tasks" }} />
+      <Stack.Screen name="create" options={{ title: "New Task" }} />
+    </Stack>
+  );
+}
