+++ apps/native/app/(tabs)/scheduler/tasks/create.tsx
@@ -1,0 +1,5 @@
+import { SchedulerTaskCreateScreen } from "@/components/scheduler/scheduler-task-create-screen";
+
+export default function SchedulerTaskCreateRoute() {
+  return <SchedulerTaskCreateScreen />;
+}
