+++ apps/native/app/(tabs)/scheduler/tasks/index.tsx
@@ -1,0 +1,5 @@
+import { SchedulerTasksScreen } from "@/components/scheduler/scheduler-tasks-screen";
+
+export default function SchedulerTasksRoute() {
+  return <SchedulerTasksScreen />;
+}
