+++ apps/native/lib/scheduler/index.ts
@@ -4,3 +4,4 @@ export * from "./date";
 export * from "./helpers";
 export * from "./notifications";
 export * from "./service";
+export * from "./use-scheduler-data";
